Build (#46)
* Changed UI in Reminder Scree

* fixed bugs and styling

* Implemented ErrorBoundary

* refactored code

* Modified-UI of addPrescriptionScreen

* commit

* Refactored navigation for Screens

* commit

* commit

* 1)refactored upload prescriptionScreen. 2)Refactored code of multiple screen

* Changed styling in Report Screen

* refactored homescreen code

* reminderScreen Refactored

* refactored code

* Added AnimatedProgressCircle

* worked on history details in reportScreen

* removed merged conflicts

* formik appointmentReminder

* refactored notificationScreen code and fixed bugs

* fixed editProfile bug

* editProfile Bug fixed

* fixed bugs and refactored code

* refactored code

* changes sendRequest Api body

* commited changes

* commited changes

* Reminder and History added Successfully

* medicine locally added successfully along with prescription

* refactored code

* refactored code

* commit

* worked on addMedicine flow

* implementing reminder feature

* commit code

* resolved merge conflicts

* Implemented Home screen progress circle

* refactored code

* commit

* appointment saved locally

+++ src/components/atoms/closeButton.js
@@ -1,35 +1,43 @@
-import { View, Text, TouchableOpacity } from 'react-native'
-import React, { useEffect } from 'react'
-import AntIcon from "react-native-vector-icons/AntDesign";
-import { colorPalette } from './colorPalette';
-import { useDispatch, useSelector } from 'react-redux';
-import { loadDeleteNotification } from '../../redux/action/notification/deleteNotificationAction';
+import {View, TouchableOpacity, Alert} from 'react-native';
+import React from 'react';
+import AntIcon from 'react-native-vector-icons/AntDesign';
+import {colorPalette} from './colorPalette';
+import {useDispatch} from 'react-redux';
+import {loadDeleteNotification} from '../../redux/action/notification/deleteNotificationAction';
+import {loadGetAllNotification} from '../../redux/action/notification/getAllNotification';
 
-const CloseButton = ({ notificationId, navigation }) => {
-  const dispatch = useDispatch()
-  const deleteNotificationdata = useSelector(state => state.deleteNotificationReducer?.data)
-  console.log(deleteNotificationdata?.status, "response")
+const CloseButton = ({notificationId}) => {
+  const dispatch = useDispatch();
 
-
-  const deleteNotification = (notificationId,) => {
-    dispatch(loadDeleteNotification(notificationId))
-    if (deleteNotificationdata?.status === 'Success') {
-    fetchAgain(navigation)
-    }
-  }
-
-  const fetchAgain=(navigation)=>{
-    navigation?.navigate('Home')
-  }
-
+  const deleteNotification = notificationId => {
+    Alert.alert('Are you sure?', 'Click ok to proceed', [
+      {
+        text: 'Ok',
+        onPress: () => {
+          dispatch(loadDeleteNotification(notificationId));
+          setTimeout(() => {
+            dispatch(loadGetAllNotification(0));
+          }, 2000);
+        },
+      },
+      {
+        text: 'Cancel',
+        onPress: () => {},
+      },
+    ]);
+  };
 
   return (
     <View>
       <TouchableOpacity onPress={() => deleteNotification(notificationId)}>
-        <AntIcon name='close' color={colorPalette.redPercentageColor} size={20} />
+        <AntIcon
+          name="close"
+          color={colorPalette.redPercentageColor}
+          size={20}
+        />
       </TouchableOpacity>
     </View>
-  )
-}
+  );
+};
 
-export default CloseButton
+export default CloseButton;
